Fixes #13
This cl fixes the hanging request problem.

This fix updates the `get` method of RequestHTML, adding logic to
retry 3 times before bailing.

The retry time I added was 5 seconds, which is a generous amount of
time... but its also a random number I pulled from hat.

This cl could be scaled down a bunch by removing the retry logic.